requirements.txt
beautifulsoup4
requests
lxml
pandas
mysql-connector-python
pyspark
